category_manager/requirements.txt
django
Pillow
toml
django-mptt
djangorestframework-simplejwt
requests